Create Animatronic change position logic
this script reads the positions available to the animatronics, calculates the position to which it will step, and it also analyzes the moves of other animatronics and if they are standing at or near the doors (8th and 6th positions), then they will not go to the occupied place. The attack frequency is also calculated.